Refactor multiplication into a module of different files
