+++ pumpkin-crates/propagators/src/propagators/arithmetic/integer_multiplication/constructor.rs
@@ -1,0 +1,70 @@
+use pumpkin_core::declare_inference_label;
+use pumpkin_core::proof::ConstraintTag;
+use pumpkin_core::proof::InferenceCode;
+use pumpkin_core::propagation::DomainEvents;
+use pumpkin_core::propagation::EventsToRegister;
+use pumpkin_core::propagation::InferenceCheckers;
+use pumpkin_core::propagation::LocalId;
+use pumpkin_core::propagation::PropagatorConstructor;
+use pumpkin_core::propagation::PropagatorConstructorContext;
+use pumpkin_core::variables::IntegerVariable;
+
+use super::checker::IntegerMultiplicationChecker;
+use super::explainer::IntegerMultiplicationExplainer;
+use super::propagator::IntegerMultiplicationPropagator;
+
+declare_inference_label!(IntegerMultiplication);
+
+const ID_A: LocalId = LocalId::from(0);
+const ID_B: LocalId = LocalId::from(1);
+const ID_C: LocalId = LocalId::from(2);
+
+/// The [`PropagatorConstructor`] for [`IntegerMultiplicationPropagator`].
+#[derive(Clone, Debug)]
+pub struct IntegerMultiplicationArgs<VA, VB, VC> {
+    pub a: VA,
+    pub b: VB,
+    pub c: VC,
+    pub constraint_tag: ConstraintTag,
+}
+
+impl<VA, VB, VC> PropagatorConstructor for IntegerMultiplicationArgs<VA, VB, VC>
+where
+    VA: IntegerVariable + 'static,
+    VB: IntegerVariable + 'static,
+    VC: IntegerVariable + 'static,
+{
+    type PropagatorImpl = IntegerMultiplicationPropagator<VA, VB, VC>;
+
+    fn add_inference_checkers(&self, mut checkers: InferenceCheckers<'_>) {
+        checkers.add_inference_checker(
+            InferenceCode::new(self.constraint_tag, IntegerMultiplication),
+            Box::new(IntegerMultiplicationChecker {
+                a: self.a.clone(),
+                b: self.b.clone(),
+                c: self.c.clone(),
+            }),
+        );
+    }
+
+    fn create(self, _: PropagatorConstructorContext) -> (EventsToRegister, Self::PropagatorImpl) {
+        let IntegerMultiplicationArgs {
+            a,
+            b,
+            c,
+            constraint_tag,
+        } = self;
+
+        let registration = EventsToRegister::builder()
+            .add(&a, DomainEvents::ANY_INT, ID_A)
+            .add(&b, DomainEvents::ANY_INT, ID_B)
+            .add(&c, DomainEvents::ANY_INT, ID_C)
+            .build();
+
+        let inference_code = InferenceCode::new(constraint_tag, IntegerMultiplication);
+        let explainer = IntegerMultiplicationExplainer::new(inference_code);
+        let propagator = IntegerMultiplicationPropagator::new(a, b, c, explainer);
+
+        (registration, propagator)
+    }
+}
