+++ pumpkin-crates/propagators/src/propagators/arithmetic/integer_multiplication/mod.rs
@@ -1,0 +1,9 @@
+mod checker;
+mod constructor;
+mod explainer;
+mod propagator;
+mod shared;
+
+pub use checker::IntegerMultiplicationChecker;
+pub use constructor::IntegerMultiplicationArgs;
+pub use propagator::IntegerMultiplicationPropagator;
